requirements.txt
streamlit==1.15.0
pandas==1.5.3
PyMuPDF==1.21.0
python-docx==0.8.11
spacy==3.5.0
numpy==1.24.0
pytesseract==0.3.10
Pillow==9.5.0
scikit-learn==1.1.2
sentence-transformers==2.2.0
fpdf==1.7.2


pip install -r requirements.txt
